Match entity & match result

+++ dugeun/src/main/java/com/project/dugeun/domain/blindDate/dao/MatchRepository.java
@@ -1,12 +1,13 @@
 package com.project.dugeun.domain.blindDate.dao;
 
 import com.project.dugeun.domain.blindDate.domain.Match;
+import com.project.dugeun.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.time.LocalDate;
-import java.util.Optional;
+import java.util.List;
 
-public interface MatchRepository extends JpaRepository<Match,Long> {
+public interface MatchRepository extends JpaRepository<Match, Integer> {
+
 
 
 }
